Change root relation payload convention.
Root relations have no source, but we treat all relations
the same when writing the payload. So by convention we set source
start end end position to target start position. This ensures that
the relation is indexed at this position.

(previous convention was that source start/end were identical to target
start/end, but this is slightly more space-efficient and 0-length feels like
a better "dummy" value for source start/end)