Smart default relation class.
When indexing, the default relation class is "rel", unless another
class is specified in the indexing configuration.

When searching, the default relation class is the first relation
class alphabetically that is not __tag.

So if your corpus only contains one relation class, you never need
to specify it.

If you do need to specify it, you can conveniently apply it for the
whole query using @rc=al.