Enable multiple annotated fields.
BlackLab has long had theoretical support for multiple annotated fields,
but in practice this didn't work properly. These issues have been fixed,
and BLS has a new field parameter to specify the annotated field to search,
if not the main one (which is the first one defined in the .blf.yaml
config).

BLS responses also include "tokenCounts" structures that report document and
corpus sizes for all of the annotated fields instead of just the main one.
For now this is an extra key in the respone, but API v5 will drop the older
"tokenCount"/"lengthInTokens" keys in favour of the new key "tokenCounts".

Also fixed an indexing bug that lead to all annotated fields getting the same
length in the forward index.

Squashed commit of the following:

commit 800bf0935feb537979836c59e1bbc3f976b97aa1
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Feb 6 14:20:04 2024 +0100

    pattfield -> field

commit 53eb4db5d1d83bc5b36c9ab6f6cda8abc795291e
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Feb 6 14:11:35 2024 +0100

    Update test responses.

commit e6be19cc7195af7dcbeb78dde3b74e898aabb489
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Feb 6 13:54:53 2024 +0100

    Index docLength separately per field.

commit c11ee29857b2bc7782d65b79e9fbf9535114d4b6
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Feb 1 15:52:46 2024 +0100

    WIP report tokenCounts per field.

commit 76db95fff50ca11d2ee3ad2516479800f9557d9d
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Feb 1 15:06:08 2024 +0100

    Use correct annotated field instead of always the main one.

commit ba982d6b70d6d23aab3797b016e45f96b5238111
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Feb 1 14:25:09 2024 +0100

    Tag suspicious mainAnnotatedField() calls.