Avoid counting optimization terms.
Relations (and therefore also inline tags) with attributes
are indexed twice: once with and once without the attributes.
This is to speed up queries where we don't filter on attributes.
But we shouldn't count these relations twice, so the extra indexed
term is now tagged as an optimization term and skipped while
counting.

Squashed commit of the following:

commit fb2c6eb8e326da067319ba7afa4c6e5f83066881
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Feb 7 09:28:30 2024 +0100

    Tweak.

commit 465c4c93190db19c73dbbc31e3a1030c5ef0bc9a
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Feb 7 08:54:45 2024 +0100

    Add comment.

commit 32c1d05e6caab6ffdc79b62c639ed1ef85d6983d
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Feb 6 16:06:19 2024 +0100

    Tag extra optimization terms to avoid counting them twice.