Report mainAnnotatedField. Get rid of special case "contents".
BLS now reports the mainAnnotatedField on the corpus info page.
This is the default field that will be used for searching and various
other operations if it's not overridden using the field parameter.

In a few places, an annotated field named "contents" was treated
differently, notably when deciding which field is the main annotated
field. This special case was removed, so the main annotated field is
always the first one created while indexing.