Drop arbitrary term length limit. Separate multiple attr values.
Relation terms with encoded attributes can get very long. BlackLab used
to impose a limit on term length of 1000, but there was no specific reason for this,
so it was dropped.

If an attribute is indexed with multiple values (because of processing steps like split),
it will be encoded into a term with space as a separator now. Before, there was no
separator.