Return '(no value)' in group properties if no value.
Grouping in BLS now returns a new JSON structure, the
group properties, with displayable values for each of the
properties that was grouped on.

If there is no value (i.e. group on word left for hits
at the start of the document), the string "(no value)"
is now returned in the `properties` structure (used to be
"-")