Fix matching bug when filtering.
The main problem was that in SpansFiltered, a conjunction that used
an approximation was used in the nextDoc() and advance() methods, which is
incorrect. All the other classes use conjunctions of the actual spans, DocIdSets,
etc., so we can use the conjunction in those methods, but approximations
are only useful for the two-phase iterator (whose whole point it is to quickly
find possibly matching docs).

Squashed commit of the following:

commit 931ebb7bc14185e0a7084da5f24d96eda3d0da99
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Apr 24 15:12:51 2024 +0200

    Fix small bug.

commit 1496d1820e2c166d8e3f9dd215aad6bfac275377
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Apr 24 15:01:19 2024 +0200

    Matching bug fixed.
    (rest of msg see above)

commit 063ed0019bd21fdc528a1d8d66d5878f745fa16d
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Apr 24 12:44:36 2024 +0200

    WIP workaround for matching bug.

    Two-phase iterator problem, temporarily disabled for now.