Parallel hits get before/match/after, not just match.
Parallel hit is determined by the span covering all relation
targets pointing to a field. So it may not be the same number
of words as in the source field, and it may also cover words
that are not directly involved in the relations matched (in between
the matching parts).

Kwic.left/right were also deprecated; use before/after instead.