Fix two-phase iteration bugs.
These bugs all stem from the fact that two-phase iteration is a system
that runs in parallel to the regular matching, and the two need to be very
careful no to mess up each other's assumptions about their clauses.

Mainly, sometimes a spans was already positioned at the first hit in a document
even though hit iteration hadn't started yet. Similarly, with bucketized hits
we sometimes fetch the first bucket as soon as we get to a document.

This causes all kinds of problems, so we now defer these hit/bucket advancements
to the latest possible time we actually need them, when we're actually finding
individual hits.