Include corpora shared with current user on server info page (/).
You can recognize these corpora because they will have an extra key,
"owner": "user <userid>".

We may add this key for all corpora in the future, probably as
"owner": "you" and "owner": "system" (for corpora that don't belong to
a specific user). So check if the "owner" key exists and start with "user "
to decide if this is a corpus shared with the current user.

Corpora shared with the current user are read-only.