Parallel: include correct "foreign hits" spans in otherFields.
Before, the foreign hits would only be determined by the target spans of
the relations matched, not by matches for the spans to the right of the ==>
operator. Now we autocapture those spans as well so the foreign hits at least
cover that. We exclude these autocaptures from the BLS response (they have
a special tag) because they are superfluous now.