rspan(...,'all') should optionally expand span. Bugfix.
Before, it would replace the current span with a span that
covers all matched relations, but what we want is actually to
adjust the span we have to include any matched relations, if
necessary. I.e. we never want to shrink our span with this call.

Also fixes a bug where the adjustment wasn't always done.