Group by match info (capture): handle unknown group name; handle missing capture.
If you try to group by a match info that doesn't exist, BLS will now respond with
a clean error message instead of an internal error.

If some hits do not have a certain match info you're grouping on, this will be handled
properly now. An empty value (no tokens) will be used in this case.