Allow BLS error response to contain structured information about error.
The "info" key is a stringmap that can contain relevant information per
error type. For now, only the error UNKNOWN_MATCH_INFO has an info key
"name" that gives the name of the unknown match info used in the query.

This allows clients to construct custom error messages that include
relevant information.