Add escapexmlfragment parameter for usecontent=orig.
usecontent=orig cuts a fragment from the original input data (instead
of reconstructing a version from the forward index). If this is not
properly escaped, it can cause issues. The escapexmlfragment will enclose
the snippet in a CDATA section.

The default value is false for API v4, but will become true starting with
API v5.

Squashed commit of the following:

commit 8f95719a7a1a983da1896321a7ff4da251627f31
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Fri Apr 5 13:18:42 2024 +0200

    Rename parameter to escapexmlfragment.

commit 0b7ae92efa5dea96f4f28284b762ee55daed71df
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Fri Apr 5 13:13:59 2024 +0200

    Allow XML fragments from usecontent=orig to be escaped as CDATA.

    In API v4 this feature must be requested explicitly using xmlfragmentcdata=true.

    Starting from API v5 (e.g. the /corpora/... URLs), this feature is enabled by
    default and must be explicitly disabled if desired (for backward compatibility).