Parallel: report matchInfo targetField, otherFields.
In the summary.pattern section of the response, matchInfos will
now include targetField for (lists of) relations, if the targetField
is different than the main search field.

The same section will also include otherFields, a list of annotated
field names that indicate the other parallel verions that were searched,
if applicable.

Neither of these will be present in non-parallel corpora.