FileProcessor: don't read archive files to byte[] if single-threaded.
FileProcessor reads files from archives into byte[] by default, because
it assumes we'll convert the files in multiple threads, and the InputStream
might not be available by the time the file is handled.

However, if we're only using 1 thread, we can use the InputStream directly.
This means we can read it into a char[] for the Saxon indexer for example,
saving the overhead of an InputStreamReader.