FileReference: prefer byte[] over char[] to use less memory.
char[] often uses about twice as much memory. On the other hand,
with byte[], the character decoding overhead applies each time the
file is processed, e.g. when parsing XML, and when retrieving the
document contents to store in the content store.

Right now, we've chosen to prefer byte[] as memory is often a
bottleneck during indexing.