Parallel: BCQL rfield(), searchfield param.
Adds rfield(..., fieldName) to BCQL to get the hits from a
specific parallel field/version. This is useful to e.g. highlight
one of the versions with the hits from a parallel query.

Adds searchfield parameter to document contents operation,
to indicate that a different field should be used as the
main (source) query field. In this case, the query should use the
above rfield() function to make sure the resulting hits are
in the correct field.

E.g. /docs/ID/contents?field=nl&searchfield=en&patt=rfield("dog" ==>nl _, 'nl')

Above example will show the Dutch version of a document with translations
for the word "dog" highlighted. This is done with a query that will search the
en field for dog and find alignments in the nl field, then will only return those
nl hits.

Squashed commit of the following:

commit a51bba3e56abc5a4ac299b8e6b65b44f261b6aa0
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 19 15:28:28 2024 +0200

    searchfield param. rfield() fixes.

commit d66789071f410ae279bbc8ee49a6cadbfc00a4b0
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Aug 19 12:39:58 2024 +0200

    Small changes to CharPosTrackingReader.

commit 8e2dd4e3af1161533b71501db63942953f7a2b43
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Fri Aug 16 12:17:38 2024 +0200

    Bugfixes, rfield() works.

commit 13ecd26eceae1e6a2b667430e5745d5b9d7599f5
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Aug 15 15:39:45 2024 +0200

    WIP rfield() function to get hits in other fields.